add sysPlot arg to plotSys in PVsystem
* if sysPlot is a figure, then clear that figure, make it current and
use it to plot the new plot
* this allows the reuse of the figure, so that the old figure doesn't
have to be closed, which is "faster" (?)
* reorganize imports in pvsystem (automatically done by pydev)
* import matplotlib.figure so that can check against Figure class, is
this really needed? How else could I check this, or should I check it at
all?
* If not using new pvsystem.PVsystem.plotSys(sysPlot) then need to close
old figure, so import close from pyplot, close self.pvapp.sysPlot and
set the figure in FigureCanvas_TkAgg to the new figure, **but** not
doing it this way, _see above_
* instead of closing old figure, call PVsystem.sysPlot with the old
figure, and just reuse it! supposedly faster?
* pvmodule_tk: comment out unused imports and disable pylint warning on
splat
* pvstring_tk: comment out unused imports and code