fix bugs in parallel_calcs
TODO: don't need to change iterable to list, numpy array is iterable!!!
* forgot to add chunksize to map of interpMods func for Vstrings
* was _not_ reshaping Vstring correctly, forgot 3rd dim, 2*NPTS
* forgot axis to sum over for Vstring! otherwise sums all elements.
* needed to expand Vsys for all strings, since there is only 1 system
* still neeeded few more steps in calc of Isys, i.e. sum over strings and
reshape to be broadcastable with Vsys when they are multiplied for Psys!
* Yay! it works!
TODO: calcMod in PVmodules is next most important loop to close, since it
is called by setSuns, which is used extensively in PVEnergy and
PVShadeLoss
TODO: calcString, low priority, is it ever used? should be able to recycle
code from parallel_calcSystem (reuse code, but make sure it is
pickle-able!, ha, ha!)