optimize parallel_calcSystem for only 1 string or 1 module
* add lots of nots
* use if statement and don't use pool if only one string or only one
 module
* flatten may be better than squeeze for interp args
* interp args are array-like, so don't need to cast as ndarray
* squeeze, repeat and sum all take array-liek so don't need to cast aas
 ndarray
FIXME: can't change class attributes in a thread because it is just a copy
(a pickle actually!) of the original class! use cls_inst.__hash__() to
print out hashes of original classes and threaded ones and you'll see they
are not the same! Python passes information between threads as pickles,
and there are no weakrefs, so the class handle is lost.